Dataset simulation

These functions allow to generate a dataset with linear dependance between \(Y\) and \(X\). data.create is used for quantitative response while data.cl.create is used for qualitative response.

Usage

data.create(n = 40, p = 10, q = 1, seed = NULL, list = TRUE)
data.cl.create(n = 40, p = 10, classes = 2, seed = NULL, list = TRUE)

Arguments

n

Number of dataset rows.

p

Number of \(X\) variables.

q

Number of \(Y\) variables with data.create function.

classes

Number of classes to generate with data.cl.create function

seed

Uses set.seed function to generate data.

list

By default, returns a list including the dataframe. If the argument list is set to FALSE, a dataframe is returned.

Details

By default, the population is set to \(n=40\) which is close to actual conditions. In this case, we have \(p<n\).

\(Y\) is a linear combinaison from each gaussian variable \(X_j\) of \(X\). Indeed, the function includes a matrix product to compute the response : \(Y = XB+E\) with \(B\) the weight (coefficients) matrix and \(E\) matrix the gaussian noise. \(B\) matrix can be found in the list returned by the function (if list = TRUE).

With data.cl.create function, Y object returned is a factor class.

Examples

library(sgPLSdevelop, warn.conflicts = FALSE, verbose = FALSE, quietly = TRUE)

# data.create
data <- data.create(n = 20, p = 5, q = 2, list = TRUE)
X <- data$X
Y <- data$Y

# data.cl.create
data.cl <- data.cl.create(n = 20, p = 5, classes = 3, list = TRUE)
X <- data.cl$X
Y <- data.cl$Y