Population projection on others dimension

This function displays scatterplot of the population (dataset individuals) projected on dimensions selected in parameters.

Usage

plot.indiv(object, compX = c(1,2), compY = NULL, ellipse.plsda = TRUE)

Arguments

object

Object of class inheriting from "sPLS", "gPLS", "sgPLS", "sPLSda", "gPLSda" or "sgPLSda". The function will retrieve some key parameters stored in that object.

compX

The X components on what the population is projected.

compY

The Y components on what the population is projected (available only for regression).

ellipse.plsda

Displays ellipses circling each point group if ellipse.plsda = TRUE (available only for classification)

Value

data

gives the latentes variables selected in parameters (in other words, the point coordinates).

graphX

scatterplot for X selected components.

graphY

scatterplot for Y selected components for regression case.

Author

Daniel FLORES

See also

Examples

library(sgPLSdevelop)

ncomp.max <- 10

# PLS 
data <- data.create(n = 20, p = 5, q = 2)
X <- data$X
Y <- data$Y
model.pls <- PLS(X,Y,ncomp = ncomp.max)
pop.model.pls <- plot.indiv(model.pls, compX = c(5,4), compY = c(1,2))

# PLSDA
data.cl <- data.cl.create(n = 20, p = 5, classes = 3)
X <- data.cl$X
Y <- data.cl$Y
model.plsda <- PLSda(X,Y,ncomp = ncomp.max)
pop.model.plsda <- plot.indiv(model.plsda, compX = c(5,4))