Plots a cluster image mapping of correlations between outcomes and all predictors

The plotcim function plots a cluster image mapping of correlations between outcomes and all the predictors.

Usage

plotcim(matX, matY, cexCol = 0.5, cexRow = 1)

Arguments

matX

data frame corresponding to the predictors.

matY

data frame corresponding to the outcomes.

cexRow, cexCol

positive numbers, used as cex.axis in for the row or column axis labeling. The defaults currently only use number of rows or columns, respectively.

Details

To be used with a small number of predictors (<1,000).

Author

Benoit Liquet, b.liquet@uq.edu.au,
Pierre Lafaye de Micheaux lafaye@dms.umontreal.ca