Group and Sparse Group Partial Least Square Model

The sgPLS package provides sparse, group and sparse group version for PLS approaches. The main functions are: sPLS for sparse PLS, gPLS for group PLS and sgPLS for sparse group PLS.

Author

Benoit Liquet <b.liquet@uq.edu.au>, Pierre Lafaye de Micheaux

References

Liquet Benoit, Lafaye de Micheaux Pierre, Hejblum Boris, Thiebaut Rodolphe. A group and Sparse Group Partial Least Square approach applied in Genomics context. Submitted.

See also